+++ nextgen/task-builder/src/tasks_builder.rs
@@ -402,8 +402,6 @@ impl<'proj> TasksBuilder<'proj> {
             };
         }
 
-        task.target = target;
-
         task.type_of = if !task.outputs.is_empty() {
             TaskType::Build
         } else if is_local {
@@ -420,9 +418,16 @@ impl<'proj> TasksBuilder<'proj> {
 
             // If an arg contains a glob, we must run in a shell for expansion to work
             if task.args.iter().any(|a| is_glob_like(a)) {
+                trace!(
+                    target = target.as_str(),
+                    "Task has a glob-like argument, wrapping in a shell so glob expansion works",
+                );
+
                 task.options.shell = Some(true);
             }
         }
+
+        task.target = target;
 
         Ok(task)
     }