+++ packages/cli/CHANGELOG.md
@@ -31,6 +31,8 @@
   per task.
 - Added environment variable support to `fileGroups`.
 - Added a `@envs(group)` token function for referencing environment variables.
+- Updated tasks with glob-like arguments to automatically enabled the `shell` option, so that glob
+  expansion works correctly.
 - Implemented a new console layer for writing to stdout/stderr.
   - Logs are now buffered and written periodically.
   - Previously they were written immediately, which required locking std each call.